+++ tasks/copy_files_to_archive/copy_files_to_archive.py
@@ -4,29 +4,48 @@ Name: copy_files_to_archive.py
 Description:  Lambda function that copies files from one s3 bucket
 to another s3 bucket.
 """
-
+import json
+import logging
 import os
 import time
-import logging
-from typing import Any, List, Dict, Optional
+from datetime import datetime, timezone
+from enum import Enum
+from typing import Any, List, Dict, Optional, Union
+import fastjsonschema
 
 import boto3
+# noinspection PyPackageRequirements
 from botocore.client import BaseClient
+# noinspection PyPackageRequirements
 from botocore.exceptions import ClientError
-import requests_db
+
+
+# todo: Move to shared lib after ORCA-170
+class RequestMethod(Enum):
+    POST = 'post'
+    PUT = 'put'
+
+
+OS_ENVIRON_DB_QUEUE_URL_KEY = 'DB_QUEUE_URL'
 
 # These will determine what the output looks like.
 FILE_SUCCESS_KEY = 'success'
-FILE_SOURCE_KEY_KEY = 'source_key'
 FILE_ERROR_MESSAGE_KEY = 'err_msg'
-FILE_REQUEST_ID_KEY = 'request_id'
-FILE_SOURCE_BUCKET_KEY = 'source_bucket'
-FILE_TARGET_BUCKET_KEY = 'target_bucket'
 
-# These are tied to the database schema.
-JOB_REQUEST_ID_KEY = 'request_id'
-JOB_ARCHIVE_BUCKET_DEST_KEY = 'archive_bucket_dest'
-JOB_JOB_STATUS_KEY = 'job_status'
+# These are tied to the input schema.
+INPUT_JOB_ID_KEY = 'job_id'
+INPUT_GRANULE_ID_KEY = 'granule_id'
+INPUT_FILENAME_KEY = 'filename'
+INPUT_SOURCE_KEY_KEY = 'source_key'
+INPUT_TARGET_KEY_KEY = 'target_key'
+INPUT_TARGET_BUCKET_KEY = 'restore_destination'
+INPUT_SOURCE_BUCKET_KEY = 'source_bucket'
+
+# todo: Move to shared lib after ORCA-170
+ORCA_STATUS_PENDING = 1
+ORCA_STATUS_STAGED = 2
+ORCA_STATUS_SUCCESS = 3
+ORCA_STATUS_FAILED = 4
 
 
 class CopyRequestError(Exception):
@@ -35,7 +54,7 @@ class CopyRequestError(Exception):
     """
 
 
-def task(records: List[Dict[str, Any]], retries: int, retry_sleep_secs: float) -> List[Dict[str, Any]]:
+def task(records: List[Dict[str, Any]], max_retries: int, retry_sleep_secs: float, db_queue_url: str) -> None:
     """
     Task called by the handler to perform the work.
 
@@ -45,156 +64,129 @@ def task(records: List[Dict[str, Any]], retries: int, retry_sleep_secs: float) -
 
     Args:
         records: Passed through from the handler.
-        retries: The number of attempts to retry a failed copy.
+        max_retries: The number of attempts to retry a failed copy.
         retry_sleep_secs: The number of seconds
             to sleep between retry attempts.
-
-    Returns:
-       A list of dicts with the following keys:
-            'source_key' (string): The object key of the file that was restored.
-            'source_bucket' (string): The name of the s3 bucket where the restored
-                file was temporarily sitting.
-            'target_bucket' (string): The name of the archive s3 bucket.
-            'success' (boolean): True, if the copy was successful,
-                otherwise False.
-            'err_msg' (string): when success is False, this will contain
-                the error message from the copy error.
-            'request_id' (string): The request_id of the database entry.
-                Only guaranteed to be present if 'success' == True.
-
-        Example:  [{'source_key': 'file1.xml', 'source_bucket': 'my-dr-fake-glacier-bucket',
-                      'target_bucket': 'unittest_xml_bucket', 'success': True,
-                      'err_msg': ''},
-                  {'source_key': 'file2.txt', 'source_bucket': 'my-dr-fake-glacier-bucket',
-                      'target_bucket': 'unittest_txt_bucket', 'success': True,
-                      'err_msg': '', 'request_id': '4192bff0-e1e0-43ce-a4db-912808c32493'}]
+        db_queue_url: The URL of the queue that posts status entries.
     Raises:
         CopyRequestError: Thrown if there are errors with the input records or the copy failed.
     """
     files = get_files_from_records(records)
-    attempt = 1
     s3 = boto3.client('s3')  # pylint: disable-msg=invalid-name
-    while attempt <= retries + 1:
+    for attempt in range(1, max_retries + 1):
         for a_file in files:
             # All files from get_files_from_records start with 'success' == False.
             if not a_file[FILE_SUCCESS_KEY]:
-                key = a_file[FILE_SOURCE_KEY_KEY]
+                err_msg = copy_object(s3,
+                                      a_file[INPUT_SOURCE_BUCKET_KEY],
+                                      a_file[INPUT_SOURCE_KEY_KEY],
+                                      a_file[INPUT_TARGET_BUCKET_KEY],
+                                      a_file[INPUT_TARGET_KEY_KEY])
+                if err_msg is None:
+                    a_file[FILE_SUCCESS_KEY] = True
+                    now = datetime.now(timezone.utc).isoformat()
+                    post_status_for_file_to_queue(
+                        a_file[INPUT_JOB_ID_KEY], a_file[INPUT_GRANULE_ID_KEY], a_file[INPUT_FILENAME_KEY], None,
+                        None,
+                        ORCA_STATUS_SUCCESS, None,
+                        None, now, now, RequestMethod.PUT, db_queue_url,
+                        max_retries, retry_sleep_secs
+                    )
+                else:
+                    a_file[FILE_ERROR_MESSAGE_KEY] = err_msg
 
-                try:
-                    if FILE_REQUEST_ID_KEY not in a_file:  # Lazily get/set the request_id
-                        job = find_job_in_db(key)
-                        if job:
-                            a_file[FILE_REQUEST_ID_KEY] = job[JOB_REQUEST_ID_KEY]
-                            a_file[FILE_TARGET_BUCKET_KEY] = job[JOB_ARCHIVE_BUCKET_DEST_KEY]
-                        else:
-                            continue
-                    err_msg = copy_object(s3, a_file[FILE_SOURCE_BUCKET_KEY], a_file[FILE_SOURCE_KEY_KEY],
-                                          a_file[FILE_TARGET_BUCKET_KEY])
-                    update_status_in_db(a_file, attempt, err_msg)
-                except requests_db.DatabaseError:
-                    continue  # Move on to the next file. We'll come back to retry on next attempt
-
-        attempt = attempt + 1
-        if attempt <= retries + 1:
+        if attempt < max_retries + 1:  # Only sleep if not on the last attempt.
             if all(a_file[FILE_SUCCESS_KEY] for a_file in files):  # Check for early completion
-                return files
+                break
             time.sleep(retry_sleep_secs)
 
-    return files
+    any_error = False
+    for a_file in files:
+        if not a_file[FILE_SUCCESS_KEY]:
+            any_error = True
+            now = datetime.now(timezone.utc).isoformat()
+            post_status_for_file_to_queue(
+                a_file[INPUT_JOB_ID_KEY], a_file[INPUT_GRANULE_ID_KEY], a_file[INPUT_FILENAME_KEY], None,
+                None,
+                ORCA_STATUS_FAILED, a_file.get(FILE_ERROR_MESSAGE_KEY, None),
+                None, now, now, RequestMethod.PUT, db_queue_url,
+                max_retries, retry_sleep_secs)
+
+    if any_error:
+        logging.error(
+            f'File copy failed. {files}')
+        raise CopyRequestError(f'File copy failed. {files}')
 
 
-def find_job_in_db(key: str) -> Optional[Dict[str, Any]]:
-    """
-    Finds the active job for the file in the database.
+# todo: Move to shared lib after ORCA-170
+def post_status_for_file_to_queue(job_id: str, granule_id: str, filename: str, key_path: Optional[str],
+                                  restore_destination: Optional[str],
+                                  status_id: Optional[int], error_message: Optional[str],
+                                  request_time: Optional[str], last_update: str,
+                                  completion_time: Optional[str],
+                                  request_method: RequestMethod,
+                                  db_queue_url: str,
+                                  max_retries: int, retry_sleep_secs: float):
+    new_data = {'job_id': job_id,
+                'granule_id': granule_id,
+                'filename': filename}
+    if key_path is not None:
+        new_data['key_path'] = key_path
+    if restore_destination is not None:
+        new_data['restore_destination'] = restore_destination
+    if status_id is not None:
+        new_data['status_id'] = status_id
+    if error_message is not None:
+        new_data['error_message'] = error_message
+    if request_time is not None:
+        new_data['request_time'] = request_time
+    if last_update is not None:
+        new_data['last_update'] = last_update
+    if completion_time is not None:
+        new_data['completion_time'] = completion_time
 
-    Args:
-        key: The object key for the file to find in the db.
-
-    Returns:
-        None if no in-progress job found for the given {key}.
-        Otherwise, the job related to the restore request. Contains the following keys:
-            'request_id' (string): The request_id of the database entry.
-            'archive_bucket_dest' (string): The name of the archive s3 bucket.
-
-    Raises:
-        requests_db.DatabaseError: Thrown if the request cannot be read from database.
-    """
-    try:
-        jobs = requests_db.get_jobs_by_object_key(key)
-    except requests_db.DatabaseError as err:
-        logging.error(f"Failed to read request from database. "
-                      f"key: {key} "
-                      f"Err: {str(err)}")
-        raise
-
-    for job in jobs:
-        if job[JOB_JOB_STATUS_KEY] != "complete":
-            return job
-
-    log_msg = ("Failed to update request status in database. "
-               f"No incomplete entry found for object_key: {key}")
-    logging.error(log_msg)
-    return None
-
-
-def update_status_in_db(a_file: Dict[str, Any], attempt: int, err_msg: str) -> None:
-    """
-    Updates the status for the job in the database.
-
-    Args:
-        a_file: An input dict with keys for:
-            'request_id' (string): The request_id of the database entry to update.
-            'source_key' (string): The filename of the restored file.
-            'source_bucket' (string): The location of the restored file.
-            'target_bucket' (string): the archive bucket the file was copied to.
-        attempt: The attempt number for the copy (1 based).
-        err_msg: None, or the error message from the copy command.
-
-    Returns:
-        None: {a_file} will be modified with additional keys for:
-            'success' (boolean): True if the copy was successful,
-                otherwise False.
-            'err_msg' (string): When 'success' is False, this will contain
-                the error message from the copy error
-
-        Example:  [{'request_id': '', 'source_key': 'file1.xml',
-                      'source_bucket': 'my-dr-fake-glacier-bucket',
-                      'target_bucket': 'unittest_xml_bucket', 'success': True,
-                      'err_msg': ''},
-                  {'request_id': '', 'source_key': 'file2.txt',
-                      'source_bucket': 'my-dr-fake-glacier-bucket',
-                      'target_bucket': 'unittest_txt_bucket', 'success': True,
-                      'err_msg': ''}]
-
-    Raises:
-        requests_db.DatabaseError: Thrown if the update operation fails.
-    """
-    new_status = ""
-    try:
-        if err_msg:
-            a_file[FILE_ERROR_MESSAGE_KEY] = err_msg
-            new_status = "error"
-            logging.error(f"Attempt {attempt}. Error copying file {a_file[FILE_SOURCE_KEY_KEY]}"
-                          f" from {a_file[FILE_SOURCE_BUCKET_KEY]} to {a_file[FILE_TARGET_BUCKET_KEY]}."
-                          f" msg: {err_msg}")
-            requests_db.update_request_status_for_job(a_file[FILE_REQUEST_ID_KEY], new_status, err_msg)
-        else:
-            a_file[FILE_SUCCESS_KEY] = True
-            a_file[FILE_ERROR_MESSAGE_KEY] = ''
-            new_status = "complete"
-            logging.info(f"Attempt {attempt}. Success copying file "
-                         f"{a_file[FILE_SOURCE_KEY_KEY]} from {a_file[FILE_SOURCE_BUCKET_KEY]} "
-                         f"to {a_file[FILE_TARGET_BUCKET_KEY]}.")
-            requests_db.update_request_status_for_job(a_file[FILE_REQUEST_ID_KEY], new_status)
-    except requests_db.DatabaseError as err:
-        logging.error(f"Failed to update request status in database. "
-                      f"key: {a_file[FILE_SOURCE_KEY_KEY]} "
-                      f"new status: {new_status}. Err: {str(err)}")
-        raise
-    return
+    post_entry_to_queue('orca_recoverfile',
+                        new_data,
+                        request_method,
+                        db_queue_url, max_retries, retry_sleep_secs)
 
 
-def get_files_from_records(records: List[Dict[str, Any]]) -> List[Dict[str, str]]:
+sqs = boto3.client('sqs')
+
+
+# todo: Move to shared lib after ORCA-170
+def post_entry_to_queue(table_name: str, new_data: Dict[str, Any], request_method: RequestMethod, db_queue_url: str,
+                        max_retries: int, retry_sleep_secs: float):
+    body = json.dumps(new_data, indent=4)
+    for attempt in range(1, max_retries + 1):
+        try:
+            sqs.send_message(
+                QueueUrl=db_queue_url,
+                MessageDeduplicationId=table_name + request_method.value + body,
+                MessageGroupId='copy_files_to_archive',
+                MessageAttributes={
+                    'RequestMethod': {
+                        'DataType': 'String',
+                        'StringValue': request_method.value
+                    },
+                    'TableName': {
+                        'DataType': 'String',
+                        'StringValue': table_name
+                    }
+                },
+                MessageBody=body
+            )
+            return
+        except Exception as e:
+            if attempt == max_retries + 1:
+                logging.error(f"Error while logging row {json.dumps(new_data, indent=4)} "
+                              f"to table {table_name}: {e}")
+                raise e
+            time.sleep(retry_sleep_secs)
+            continue
+
+
+def get_files_from_records(records: List[Dict[str, Any]]) -> List[Dict[str, Union[str, bool]]]:
     """
     Parses the input records and returns the files to be restored.
 
@@ -202,29 +194,18 @@ def get_files_from_records(records: List[Dict[str, Any]]) -> List[Dict[str, str]
         records: passed through from the handler.
 
     Returns:
-        files: A list of dicts with the following keys:
-            'source_key': The object key of the file that was restored.
-            'source_bucket': The name of the s3 bucket where the restored
-                file was temporarily sitting.
-
-    Raises:
-        KeyError: Thrown if the event record does not contain a value for the given bucket or object.
-            Bucket will be checked/raised first, then object.
+        records, parsed into Dicts, with the additional KVP 'success' = False
     """
+    with open("schemas/sub_schemas/body.json", "r") as raw_schema:
+        schema = json.loads(raw_schema.read())
+
+    validate = fastjsonschema.compile(schema)
     files = []
     for record in records:
-        a_file = {FILE_SUCCESS_KEY: False}
-        log_str = 'Records["s3"]["bucket"]["name"]'
-        try:
-            source_bucket = record["s3"]["bucket"]["name"]
-            a_file[FILE_SOURCE_BUCKET_KEY] = source_bucket
-            log_str = 'Records["s3"]["object"]["key"]'  # Use the same var to make except handling easier.
-            source_key = record["s3"]["object"]["key"]
-            a_file[FILE_SOURCE_KEY_KEY] = source_key
-            files.append(a_file)
-        except KeyError:
-            raise CopyRequestError(f'event record: "{record}" does not contain a '
-                                   f'value for {log_str}')
+        a_file = json.loads(record['body'])
+        validate(a_file)
+        a_file[FILE_SUCCESS_KEY] = False
+        files.append(a_file)
     return files
 
 
@@ -258,11 +239,12 @@ def copy_object(s3_cli: BaseClient, src_bucket_name: str, src_object_name: str,
         logging.debug(f"Copy response: {response}")
     except ClientError as ex:
         logging.error(ex)
-        return str(ex)
+        return ex.__str__()
     return None
 
 
-def handler(event: Dict[str, Any], context: object) -> List[Dict[str, Any]]:  # pylint: disable-msg=unused-argument
+# noinspection PyUnusedLocal
+def handler(event: Dict[str, Any], context: object) -> None:  # pylint: disable-msg=unused-argument
     """Lambda handler. Copies a file from its temporary s3 bucket to the s3 archive.
 
     If the copy for a file in the request fails, the lambda
@@ -283,66 +265,16 @@ def handler(event: Dict[str, Any], context: object) -> List[Dict[str, Any]]:  # 
                 drdb-host (string): the database host
 
     Args:
-        event: A dict with the following keys:
-
-            Records (list(dict)): A list of dict with the following keys:
-                s3 (dict): A dict with the following keys:
-                    bucket (dict):  A dict with the following keys:
-                        name (string): The name of the s3 bucket holding the restored file
-                    object (dict):  A dict with the following keys:
-                        key (string): The key of the restored file
-
-            Example: event: {"Records": [{"eventVersion": "2.1",
-                                  "eventSource": "aws:s3",
-                                  "awsRegion": "us-west-2",
-                                  "eventTime": "2019-06-17T18:54:06.686Z",
-                                  "eventName": "ObjectRestore:Post",
-                                  "userIdentity": {
-                                  "principalId": "AWS:AROAJWMHUPO:request_files"},
-                                  "requestParameters": {"sourceIPAddress": "1.001.001.001"},
-                                  "responseElements": {"x-amz-request-id": "0364DB32C0",
-                                                       "x-amz-id-2":
-                                     "4TpisFevIyonOLD/z1OGUE/Ee3w/Et+pr7c5F2RbnAnU="},
-                                  "s3": {"s3SchemaVersion": "1.0",
-                                        "configurationId": "dr_restore_complete",
-                                        "bucket": {"name": exp_src_bucket,
-                                                   "ownerIdentity":
-                                                   {"principalId": "A1BCXDGCJ9"},
-                                           "arn": "arn:aws:s3:::my-dr-fake-glacier-bucket"},
-                                        "object": {"key": exp_file_key1,
-                                                   "size": 645,
-                                                   "sequencer": "005C54A126FB"}}}]}
+        event:
+            A dict from the SQS queue. See schemas/input.json for more information.
 
         context: An object required by AWS Lambda. Unused.
-
-    Returns:
-        The list of dicts returned from the task. All 'success' values will be True. If they were
-        not all True, the CopyRequestError exception would be raised.
-        Dicts have the following keys:
-            'source_key' (string): The object key of the file that was restored.
-            'source_bucket' (string): The name of the s3 bucket where the restored
-                file was temporarily sitting.
-            'target_bucket' (string): The name of the archive s3 bucket
-            'success' (boolean): True, if the copy was successful,
-                otherwise False.
-            'err_msg' (string): when success is False, this will contain
-                the error message from the copy error.
-            'request_id' (string): The request_id of the database entry.
-                Only guaranteed to be present if 'success' == True.
-
-        Example:  [{'source_key': 'file1.xml', 'source_bucket': 'my-dr-fake-glacier-bucket',
-                      'target_bucket': 'unittest_xml_bucket', 'success': True,
-                      'err_msg': ''},
-                  {'source_key': 'file2.txt', 'source_bucket': 'my-dr-fake-glacier-bucket',
-                      'target_bucket': 'unittest_txt_bucket', 'success': True,
-                      'err_msg': '', 'request_id': '4192bff0-e1e0-43ce-a4db-912808c32493'}]
 
     Raises:
         CopyRequestError: An error occurred calling copy_object for one or more files.
         The same dict that is returned for a successful copy will be included in the
         message, with 'success' = False for the files for which the copy failed.
     """
-
     logging.basicConfig(level=logging.INFO,
                         format='%(levelname)s: %(asctime)s: %(message)s')
     try:
@@ -357,13 +289,12 @@ def handler(event: Dict[str, Any], context: object) -> List[Dict[str, Any]]:  # 
     except KeyError:
         retry_sleep_secs = 30
 
+    try:
+        db_queue_url = str(os.environ[OS_ENVIRON_DB_QUEUE_URL_KEY])
+    except KeyError as key_error:
+        logging.error(f"os.environ[{OS_ENVIRON_DB_QUEUE_URL_KEY}] not found.")
+        raise key_error
     logging.debug(f'event: {event}')
     records = event["Records"]
-    result = task(records, retries, retry_sleep_secs)
-    for a_file in result:
-        # if any file failed, the function will fail
-        if not a_file[FILE_SUCCESS_KEY]:
-            logging.error(
-                f'File copy failed. {result}')
-            raise CopyRequestError(f'File copy failed. {result}')
-    return result
+
+    task(records, retries, retry_sleep_secs, db_queue_url)