+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and includes an additional section for migration notes.
 ## [Unreleased]
 
 ### Changed
+- *ORCA-109* request_files now uses SQS queue for recovery status updates, and receives input from a separate SQS queue.
 - *ORCA-91* copy_files_to_archive now uses SQS queue for recovery status updates. Will generate a job_id if none is given, and return it in the output.
 - request_files now uses the same default glacier bucket as copy_to_glacier.